requirements_appveyor.txt
xlrd>=1.0.0
matplotlib>=1.5.3
cryptography>=1.5.2
click>=6.6
lmfit>=0.9.7
setuptools>=35.0.1
pandas>=0.23.0
pyyaml>=3.12
pyodbc>=4.0.16
python-box>=3.2.0
PyGithub>=1.43
numpy==1.15.4
pytest-benchmark
pytest-timeout
